# test: Use MiniWallet in rpc_rawtransaction.py