@danielabrozzoni 2022-05-30
Thanks for your review @kouloumos! You're right, the 4th node is useless - I updated the code to use only 3 nodes, (and also removed the useless `sync_all`), and it runs slightly faster.

Before:
`test/functional/rpc_rawtransaction.py  2.12s user 0.64s system 27% cpu 10.121 total`
After:
`test/functional/rpc_rawtransaction.py  1.58s user 0.46s system 26% cpu 7.798 total`