@jonatack 2022-05-30
There is the Part 2 of https://github.com/bitcoin/bitcoin/pull/22437 (initially #24113 before it was chopped to bugfix only) that I was planning to repropose after updating for the changes here. Won't do it right away so it can be used if anyone wants to:   https://github.com/jonatack/bitcoin/commits/improve-getrawtransaction-tests-part-2